+++ src/main/java/com/axonactive/coffeeshopmanagement/service/dto/InvoiceDetailDto.java
@@ -7,8 +7,8 @@ public class InvoiceDetailDto {
 
     private String invoiceId;
     private String itemName;
-    private Double itemPrice;
-    private Double finalPrice;
+    private Integer itemPrice;
+    private Integer finalPrice;
     private Integer quantity;
     private Double discount;
